Move gometa tests to their own file
